#003 - Create Campus module
