--- src/main/java/de/fherfurt/campus/Address.java
@@ -1,13 +1,0 @@
-package de.fherfurt.campus;
-
-public class Address {
-  private String street;
-
-  public String getStreet() {
-    return street;
-  }
-
-  public void setStreet(String street) {
-    this.street = street;
-  }
-}
