+++ src/main/java/de/fherfurt/campus/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides the API to the Campus module. It allows the loading of an Location via an string or an
+ * {@link de.fherfurt.campus.transfer.objects.Address}. The current dummy-implementation will be replaced with the correct implementation
+ * of one of the modules.
+ */
+package de.fherfurt.campus;
