+++ src/main/java/de/fherfurt/campus/transfer/objects/Address.java
@@ -1,0 +1,13 @@
+package de.fherfurt.campus.transfer.objects;
+
+public class Address {
+    private String street;
+
+    public String getStreet() {
+        return street;
+    }
+
+    public void setStreet(String street) {
+        this.street = street;
+    }
+}
